Better beginning to Blackjack

+++ src/main/java/com/example/cyber_proovikas/blackjackServer/playerControl/Player.java
@@ -14,6 +14,6 @@ public class Player {
     @Id
     String username;
 
-    int game;
+    long game;
     BigDecimal funds;
 }
